MXGDA-3723 Minor refactor from review
